+++ tests/test_app_legacy.py
@@ -142,63 +142,3 @@ def test_app_ocf_datapipes(
         == expected_forecast_results * expected_forecast_timesteps
     )
 
-# Test for new DA model with data sampler utilisation
-# To note - Satellite omitted
-def test_app_day_ahead_data_sampler(
-    db_session, nwp_ukv_data, nwp_ecmwf_data, gsp_yields_and_systems, me_latest
-):
-    """Test the app running the day ahead model"""
-
-    with tempfile.TemporaryDirectory() as tmpdirname:
-
-        os.chdir(tmpdirname)
-
-        temp_nwp_path = "temp_nwp_ukv.zarr"
-        os.environ["NWP_UKV_ZARR_PATH"] = temp_nwp_path
-        nwp_ukv_data.to_zarr(temp_nwp_path)
-
-        temp_nwp_path = "temp_nwp_ecmwf.zarr"
-        os.environ["NWP_ECMWF_ZARR_PATH"] = temp_nwp_path
-        nwp_ecmwf_data.to_zarr(temp_nwp_path)
-
-        os.environ["SATELLITE_ZARR_PATH"] = "nonexistent_sat.zarr.zip"
-        os.environ["DAY_AHEAD_MODEL"] = "True"
-        os.environ["RUN_EXTRA_MODELS"] = "False"
-        os.environ["USE_OCF_DATA_SAMPLER"] = "True"
-
-        # Import at runtime to ensure environment variables are set
-        from pvnet_app.app import app
-        app(gsp_ids=list(range(1, 318)), num_workers=2)
-
-    all_models = get_all_models(get_day_ahead_only=True, use_ocf_data_sampler=True)
-
-    # Check correct number of forecasts have been made
-    # (317 GSPs + 1 National + maybe GSP-sum) = 318 or 319 forecasts
-    # Forecast made with multiple models
-    expected_forecast_results = 0
-    for model_config in all_models:
-        expected_forecast_results += 318 + model_config.save_gsp_sum
-
-    forecasts = db_session.query(ForecastSQL).all()
-    # Doubled for historic and forecast
-    assert len(forecasts) == expected_forecast_results * 2
-
-    # Check probabilistic added
-    assert "90" in forecasts[0].forecast_values[0].properties
-    assert "10" in forecasts[0].forecast_values[0].properties
-
-    # 72 time steps in forecast
-    expected_forecast_timesteps = 72
-
-    assert (
-        len(db_session.query(ForecastValueSQL).all())
-        == expected_forecast_results * expected_forecast_timesteps
-    )
-    assert (
-        len(db_session.query(ForecastValueLatestSQL).all())
-        == expected_forecast_results * expected_forecast_timesteps
-    )
-    assert (
-        len(db_session.query(ForecastValueSevenDaysSQL).all())
-        == expected_forecast_results * expected_forecast_timesteps
-    )
